Replace the size in BoxHeader with an enum
In mp4, box sizes of 0 mean that the box extends to the end of the file. But
some parts of the code don't account for this and take 0 to mean that the box
literally has a size of 0, which is impossible and causes bugs such as infinite
loops. Replace it with an enum to help ensure that the “until EOF” case is
handled correctly.